* rdno_core
* nvstore; implemenation for ESP8266
* config; storing u8, u16, u32, etc.. is now done into a byte buffer
* packet; the sensor value is now utterly fixed to u16/s16
* network; port is u16

+++ source/main/cpp/c_network.cpp
@@ -139,6 +139,6 @@ namespace ncore
         return from_string(str, addr);
     }
 
-    bool is_valid_port(s32 port) { return port > 1023 && port <= 65535; }
+    bool is_valid_port(u16 port) { return port > 1023 && port <= 65535; }
 
 }  // namespace ncore
